+++ packages/plugin-automations-ui/src/components/forms/actions/subForms/sendEmail/recipientForm.tsx
@@ -1,47 +1,229 @@
 import {
+  Button,
   Chip,
   ControlLabel,
   EmptyState,
   FormControl,
   FormGroup,
   Icon,
-  Label,
   SelectTeamMembers,
-  TabTitle,
-  Tabs,
   __,
-  colors
-} from '@erxes/ui/src';
+  colors,
+} from "@erxes/ui/src";
 
-import PlaceHolderInput from '@erxes/ui-automations/src/components/forms/actions/placeHolder/PlaceHolderInput';
-import { DrawerDetail } from '@erxes/ui-automations/src/styles';
-import SelectCompanies from '@erxes/ui-contacts/src/companies/containers/SelectCompanies';
-import SelectCustomers from '@erxes/ui-contacts/src/customers/containers/SelectCustomers';
-import Popover from '@erxes/ui/src/components/Popover';
-import { Avatar } from '@erxes/ui/src/components/SelectWithSearch';
-import { PopoverContent } from '@erxes/ui/src/components/filterableList/styles';
-import React, { useState } from 'react';
-import { renderDynamicComponent } from '../../../../../utils';
-import { Padding } from '../../styles';
-import { checkToFieldConfigured } from './utils';
-import { FieldsCombinedByType } from '@erxes/ui-forms/src/settings/properties/types';
+import PlaceHolderInput from "@erxes/ui-automations/src/components/forms/actions/placeHolder/PlaceHolderInput";
+import SelectCompanies from "@erxes/ui-contacts/src/companies/containers/SelectCompanies";
+import SelectCustomers from "@erxes/ui-contacts/src/customers/containers/SelectCustomers";
+import { FieldsCombinedByType } from "@erxes/ui-forms/src/settings/properties/types";
+import { FlexRow } from "@erxes/ui-settings/src/styles";
+import Popover from "@erxes/ui/src/components/Popover";
+import { Avatar } from "@erxes/ui/src/components/SelectWithSearch";
+import { PopoverContent } from "@erxes/ui/src/components/filterableList/styles";
+import { FlexColumn } from "@erxes/ui/src/components/step/style";
+import React, { useState } from "react";
+import Select from "react-select";
+import { renderDynamicComponent } from "../../../../../utils";
+
+const MAIL_TYPES = [
+  { type: "attributionMail", name: "attributionMails", label: "Dynamic Entry" },
+  { type: "customMail", name: "customMails", label: "Manual Entry" },
+  {
+    type: "teamMember",
+    name: "teamMemberIds",
+    label: "Team Members",
+  },
+  {
+    type: "lead",
+    name: "leadIds",
+    label: "Leads",
+  },
+  {
+    type: "customer",
+    name: "customerIds",
+    label: "Customers",
+  },
+  {
+    type: "company",
+    name: "companyIds",
+    label: "Companies",
+  },
+];
 
 export const RecipientsForm = ({
-  emailRecipientsConst,
   triggerType,
   triggerConfig,
   config,
   onChangeConfig,
-  additionalAttributes
+  additionalAttributes,
 }) => {
-  const [selectedTab, setTab] = useState('general');
+  const [CCEnabled, setCCEnabled] = useState(config?.cc || false);
 
   const onSelect = (value: any, name: string) => {
     onChangeConfig({ ...config, [name]: value });
   };
 
-  const renderToEmailsContent = () => {
-    if (selectedTab === 'general') {
+  return (
+    <>
+      <FormGroup>
+        <FlexRow $justifyContent='space-between'>
+          <ControlLabel>{__("To")}</ControlLabel>
+          <Button
+            size='small'
+            btnStyle={CCEnabled ? "white" : "link"}
+            onClick={() => {
+              if (CCEnabled) {
+                onChangeConfig({ ...config, cc: undefined });
+                setCCEnabled(false);
+              } else {
+                setCCEnabled(true);
+              }
+            }}
+          >
+            {__("CC")}
+          </Button>
+        </FlexRow>
+        <FieldRow
+          config={config}
+          onSelect={onSelect}
+          onChangeConfig={onChangeConfig}
+          additionalAttributes={additionalAttributes}
+          triggerConfig={triggerConfig}
+          triggerType={triggerType}
+        />
+      </FormGroup>
+      {CCEnabled && (
+        <FormGroup>
+          <ControlLabel>{__("CC")}</ControlLabel>
+          <FieldRow
+            config={config?.cc || {}}
+            onSelect={(value: any, name: string) =>
+              onSelect({ ...(config?.cc || {}), [name]: value }, "cc")
+            }
+            onChangeConfig={(newConfig) =>
+              onChangeConfig({
+                ...(config || {}),
+                cc: { ...(newConfig || {}) },
+              })
+            }
+            additionalAttributes={additionalAttributes}
+            triggerConfig={triggerConfig}
+            triggerType={triggerType}
+          />
+        </FormGroup>
+      )}
+    </>
+  );
+};
+
+const FieldRow = ({
+  config,
+  onSelect,
+  onChangeConfig,
+  additionalAttributes,
+  triggerType,
+  triggerConfig,
+}: {
+  config: any;
+  onSelect: (...args: any) => void;
+  triggerType: string;
+  triggerConfig: any;
+  onChangeConfig: (config: any) => void;
+  additionalAttributes: FieldsCombinedByType[];
+}) => {
+  const [selectedType, setType] = useState("attributionMail");
+  const options = MAIL_TYPES.map(({ type, label }) => ({ value: type, label }));
+
+  return (
+    <FlexColumn style={{ gap: "12px" }}>
+      <Select
+        defaultValue={options.find(({ value }) => value === selectedType)}
+        options={options}
+        onChange={(opt) =>
+          opt?.value ? setType(opt?.value) : setType("attributionMail")
+        }
+      />
+
+      <EmailFieldInputWrapper
+        type={selectedType}
+        config={config}
+        onSelect={onSelect}
+        onChangeConfig={onChangeConfig}
+        additionalAttributes={additionalAttributes}
+        triggerConfig={triggerConfig}
+        triggerType={triggerType}
+      />
+    </FlexColumn>
+  );
+};
+
+const EmailFieldInputWrapper = ({
+  type,
+  config,
+  onSelect,
+  onChangeConfig,
+  additionalAttributes,
+  triggerType,
+  triggerConfig,
+}: {
+  type: string;
+  config: any;
+  onSelect: (...args: any) => void;
+  triggerType: string;
+  triggerConfig: any;
+  onChangeConfig: (config: any) => void;
+  additionalAttributes: FieldsCombinedByType[];
+}) => {
+  switch (type) {
+    case "customMail":
+      return <CustomMailInput config={config} onSelect={onSelect} />;
+    case "teamMember":
+      return (
+        <FormGroup>
+          <ControlLabel>{__("Team member")}</ControlLabel>
+          <SelectTeamMembers
+            name={"teamMemberIds"}
+            initialValue={config["teamMemberIds"] || ""}
+            label={"Team members"}
+            onSelect={onSelect}
+            filterParams={{
+              status: "Verified",
+            }}
+          />
+        </FormGroup>
+      );
+    case "customer":
+    case "lead":
+      return (
+        <FormGroup>
+          <ControlLabel>{__(type.toLocaleUpperCase())}</ControlLabel>
+          <SelectCustomers
+            name={`${type}Ids`}
+            initialValue={config[`${type}Ids`] || ""}
+            label={`${type.toLocaleUpperCase()}s`}
+            onSelect={onSelect}
+            filterParams={{
+              status: "Verified",
+            }}
+          />
+        </FormGroup>
+      );
+    case "company":
+      return (
+        <FormGroup>
+          <ControlLabel>{__("Customer")}</ControlLabel>
+          <SelectCompanies
+            name={"companyIds"}
+            initialValue={config["companyIds"] || ""}
+            label={"Companies"}
+            onSelect={onSelect}
+            filterParams={{
+              status: "Verified",
+            }}
+          />
+        </FormGroup>
+      );
+    case "attributionMail":
+    default:
       return (
         <AttrubutionInput
           config={config}
@@ -51,64 +233,7 @@ export const RecipientsForm = ({
           additionalAttributes={additionalAttributes}
         />
       );
-    }
-
-    if (selectedTab === 'static') {
-      return (emailRecipientsConst || []).map((emailRType) => (
-        <RecipientTypeComponent
-          props={{ config, onSelect }}
-          emailRType={emailRType}
-        />
-      ));
-    }
-
-    return null;
-  };
-
-  return (
-    <FormGroup>
-      <ControlLabel>{__('To Emails')}</ControlLabel>
-      <DrawerDetail>
-        <Tabs full>
-          <TabTitle
-            className={selectedTab === 'general' ? 'active' : ''}
-            onClick={() => setTab('general')}
-          >
-            {__('General')}
-            {config?.attributionMails && (
-              <Label lblStyle='danger'>
-                <Icon icon='check' />
-              </Label>
-            )}
-          </TabTitle>
-          <TabTitle
-            className={selectedTab === 'static' ? 'active' : ''}
-            onClick={() => setTab('static')}
-          >
-            {__('Static')}
-            {checkToFieldConfigured(emailRecipientsConst, config) && (
-              <Label lblStyle='danger' shake={true} ignoreTrans={true}>
-                {`${emailRecipientsConst
-                  .filter(({ name }) => name !== 'attributionMails')
-                  .reduce((acc, item) => {
-                    if (
-                      config.hasOwnProperty(item.name) &&
-                      Array.isArray(config[item.name])
-                        ? (config[item.name] || []).length
-                        : config[item.name]
-                    ) {
-                      acc++;
-                    }
-                    return acc;
-                  }, 0)}`}
-              </Label>
-            )}
-          </TabTitle>
-        </Tabs>
-        <Padding>{renderToEmailsContent()}</Padding>
-      </DrawerDetail>
-    </FormGroup>
-  );
+  }
 };
 
 const AttrubutionInput = ({
@@ -116,7 +241,7 @@ const AttrubutionInput = ({
   triggerType,
   triggerConfig,
   additionalAttributes,
-  config
+  config,
 }: {
   config: any;
   onChangeConfig: (config: any) => void;
@@ -126,7 +251,7 @@ const AttrubutionInput = ({
 }) => {
   const onChange = (updatedConfig) => onChangeConfig(updatedConfig);
 
-  const isAvailableTriggerExecutor = ['customer', 'companies', 'user'].some(
+  const isAvailableTriggerExecutor = ["customer", "companies", "user"].some(
     (c) => triggerType.includes(c)
   );
 
@@ -134,22 +259,22 @@ const AttrubutionInput = ({
     ? [
         {
           _id: String(Math.random()),
-          label: 'Trigger Executors',
-          name: 'triggerExecutors',
-          type: 'segment'
-        }
+          label: "Trigger Executors",
+          name: "triggerExecutors",
+          type: "segment",
+        },
       ]
     : [];
 
   return (
     <PlaceHolderInput
       config={config}
-      triggerType={triggerType}
+      triggerType={triggerType.split(".")[0]}
       inputName='attributionMails'
-      placeholder={__('Please select  some attributes from attributes section')}
+      placeholder={__("Please select  some attributes from attributes section")}
       label='Dynamic mails'
-      attrTypes={['user', 'contact', 'segment']}
-      attrWithSegmentConfig={triggerType === 'forms:form_submission'}
+      attrTypes={["user", "contact", "segment", "users"]}
+      attrWithSegmentConfig={triggerType === "forms:form_submission"}
       triggerConfig={triggerConfig}
       onChange={onChange}
       customAttributions={[...additionalAttributes, ...customAttributions]}
@@ -158,7 +283,7 @@ const AttrubutionInput = ({
           placement='auto'
           trigger={<Icon color={colors.colorCoreRed} icon='question-circle' />}
         >
-          <PopoverContent style={{ width: '200px', padding: '10px' }}>
+          <PopoverContent style={{ width: "200px", padding: "10px" }}>
             This type does not include (From Mail) and (Not Verified mails)
           </PopoverContent>
         </Popover>
@@ -169,7 +294,7 @@ const AttrubutionInput = ({
 
 const CustomMailInput = ({
   config,
-  onSelect
+  onSelect,
 }: {
   config: any;
   onSelect: (value: any, name) => void;
@@ -177,24 +302,24 @@ const CustomMailInput = ({
   const onChange = (e) => {
     const { value } = e.currentTarget;
     if (
-      e.key === 'Enter' &&
+      e.key === "Enter" &&
       value.match(/^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/)
     ) {
-      onSelect((config?.customMails || []).concat(value), 'customMails');
-      e.currentTarget.value = '';
+      onSelect((config?.customMails || []).concat(value), "customMails");
+      e.currentTarget.value = "";
     }
   };
 
   const removeMail = (mail) => {
     onSelect(
       (config?.customMails || []).filter((value) => value !== mail),
-      'customMails'
+      "customMails"
     );
   };
 
   return (
     <FormGroup>
-      <ControlLabel>{__('Custom Mail')}</ControlLabel>
+      <ControlLabel>{__("Custom Mail")}</ControlLabel>
       {(config?.customMails || []).map((customMail) => (
         <Chip
           key={customMail}
@@ -211,7 +336,7 @@ const CustomMailInput = ({
 
 const RecipientTypeComponent = ({
   props: { config = {}, onSelect },
-  emailRType: { serviceName, label, name, type }
+  emailRType: { serviceName, label, name, type },
 }: {
   props: { config: any; onSelect: (value: any, name: any) => void };
   emailRType: {
@@ -224,31 +349,31 @@ const RecipientTypeComponent = ({
   if (serviceName) {
     return renderDynamicComponent(
       {
-        componentType: 'selectRecipients',
+        componentType: "selectRecipients",
         type,
         value: config[name],
         label,
         name,
-        onSelect
+        onSelect,
       },
       `${serviceName}:${type}`
     );
   }
 
   switch (type) {
-    case 'customMail':
+    case "customMail":
       return <CustomMailInput config={config} onSelect={onSelect} />;
-    case 'teamMember':
+    case "teamMember":
       return (
         <FormGroup>
           <ControlLabel>{__(label)}</ControlLabel>
           <SelectTeamMembers
             name={name}
-            initialValue={config[name] || ''}
+            initialValue={config[name] || ""}
             label={label}
             onSelect={onSelect}
             filterParams={{
-              status: 'Verified'
+              status: "Verified",
             }}
           />
         </FormGroup>
@@ -258,7 +383,7 @@ const RecipientTypeComponent = ({
       const Components = {
         lead: SelectCustomers,
         customer: SelectCustomers,
-        company: SelectCompanies
+        company: SelectCompanies,
       };
 
       const Component = Components[type];
@@ -272,12 +397,12 @@ const RecipientTypeComponent = ({
           <ControlLabel>{__(label)}</ControlLabel>
           <Component
             name={name}
-            initialValue={config[name] || ''}
+            initialValue={config[name] || ""}
             label={label}
             onSelect={onSelect}
             filterParams={{
               type,
-              emailValidationStatus: 'valid'
+              emailValidationStatus: "valid",
             }}
           />
         </FormGroup>